Create countdown_to_new_year.html
In the above code, I've added a <style> block within the <head> section. This block contains CSS rules that apply to the element with the id "countdown". The "font-size: 3em;" line increases the font size of the text within the countdown paragraph. You can adjust the "3em" value to increase or decrease the size as you see fit.

The "em" unit is a relative unit of measurement in CSS, which is relative to the current font size of the element or the parent element. 1em is equal to the current font size, so 3em means 3 times the current font size. You could also use other units like pixels (px), points (pt), or viewport width (vw) if you prefer.